Escreva um programa que pergunte a velocidade de um carro. Caso ultrapasse 80Km/h, exiba uma mensagem dizendo que o usuário foi multado. Nesse caso, exiba o valor da multa, cobrando R$5 por cada Km acima da velocidade permitida.

+++ src/Estudo_Condicionais/Atv01RadarEletronico.java
@@ -8,7 +8,7 @@ public class Atv01RadarEletronico {
         double velocidade;
 
         System.out.print("Digite a velocidade do carro: ");
-        velocidade = scan.nextInt();
+        velocidade = scan.nextDouble();
         double valorMulta = (velocidade - 80) * 5;
 
         if (velocidade > 80){
